feat(gemma3): consolidate text decoder into MLXLMCommon namespace (WS-C #1, issue #168)
Lift the Gemma 3 text-decoder layer stack into a public
`Gemma3` namespace under `MLXLMCommon/Models/Gemma3.swift`.
Both targets now consume the shared types:

  Libraries/MLXLMCommon/Models/Gemma3.swift  ←  shared
    - public enum Gemma3
    - public struct Gemma3.TextConfiguration   (handles flat + nested
                                                `text_config` JSON)
    - public class  Gemma3.Attention           (uses applyRotaryPosition;
                                                supports BatchPositionedKVCache
                                                via the helper)
    - public class  Gemma3.MLP
    - public class  Gemma3.TransformerBlock
    - public class  Gemma3.Backbone            (embed → N blocks → norm,
                                                accepts `inputs` OR
                                                `inputEmbedding` for
                                                vision-fusion)

  Libraries/MLXLLM/Models/Gemma3.swift        ← (renamed from
                                                Gemma3Text.swift)
    - thin Gemma3TextModel: LLMModel wrapper
    - back-compat typealias for Gemma3TextConfiguration

  Libraries/MLXVLM/Models/Gemma3.swift        ← stripped 350+ lines
    - drops the duplicated private Attention/MLP/TransformerBlock/
      GemmaModel
    - LanguageModel now wraps Gemma3.Backbone, adds Linear-or-
      QuantizedLinear lm_head dispatch + final-logit softcapping
    - drops the dead `hasQuantizedLmHead` quant gate (per issue #170
      cleanup — quantization is applied via the top-level
      `quantize(model:)` block in Load.swift driven by
      `BaseConfiguration.PerLayerQuantization`)
    - shared-namespace references qualified with `MLXLMCommon.` to
      disambiguate from the local `public class Gemma3` outer model
    - prepare gains the same hard `eval(cacheArrays + [result.logits])`
      barrier as Gemma 4 VLM (issue #169 fix in 4f6c9b7) — the
      pre-consolidation private `GemmaModel` happened to mask the
      prefill-sync issue; the consolidated Backbone surfaces it the
      same way Gemma 4 VLM did

Net diff: +542 / -347 lines (~200-line reduction net of the new
namespace file).

Verification (alpha @ this commit):
- `swift build` clean across MLXLMCommon, MLXLLM, MLXVLM.
- `swift test --configuration release` = 203/204 (the 1 failure is
  the env-driven InferenceBenchmark scaffold, pre-existing).
- `gemma-3-1b-it-qat-4bit --method simple` →
  "Hello! I'm Gemma..." (121 tok/s)
- `gemma-3-4b-it-qat-4bit --method simple` →
  "Hello! My name is Gemma..." (83 tok/s)
- `gemma-3-4b-it-qat-4bit --method vision` →
  "Okay, here's a description of the image:..." (89 tok/s, validation
  passes — keyword "dog" in 200-token output)
- `gemma-3-1b --method summarization 1024 --kv affine4` →
  "The..." continuation (124 tok/s)
- `gemma-3-4b --method summarization 1024 --kv affine4` →
  "Okay, here's a summary..." (42 tok/s)

Closes WS-C #1 of issue #168 sprint plan.
